Improvements to RPC parsing and startup checks
This fixes some pretty glaring issues in the RPC startup parsing and
testing.

This patch also moves the code to a location where it can be called in
the integration tests.

Finally a new check was added validating that the orchestrator address
does in fact have the fee token that is being specified to spend.
