+++ orchestrator/peggy_utils/Cargo.toml
@@ -23,3 +23,4 @@ log = "0.4"
 url = "2"
 [dev_dependencies]
 rand = "0.8"
+actix = "0.10"
