+++ orchestrator/peggy_utils/src/connection_prep.rs
@@ -2,8 +2,13 @@
 //! It's a common problem to have conflicts between ipv4 and ipv6 localhost and this module is first and foremost supposed to resolve that problem
 //! by trying more than one thing to handle potentially misconfigured inputs.
 
+use clarity::Address as EthAddress;
 use contact::client::Contact;
+use deep_space::address::Address as CosmosAddress;
 use peggy_proto::peggy::query_client::QueryClient as PeggyQueryClient;
+use peggy_proto::peggy::QueryDelegateKeysByEthAddress;
+use peggy_proto::peggy::QueryDelegateKeysByOrchestratorAddress;
+use std::process::exit;
 use std::time::Duration;
 use tonic::transport::Channel;
 use url::Url;
@@ -39,6 +44,10 @@ pub async fn create_rpc_connections(
             Ok(val) => grpc = Some(val),
             // did not work, now we check if it's localhost
             Err(e) => {
+                warn!(
+                    "Failed to access Cosmos gRPC with {:?} trying fallback options",
+                    e
+                );
                 if grpc_url.to_lowercase().contains("localhost") {
                     let port = url.port().unwrap_or(80);
                     // this should be http or https
@@ -47,6 +56,7 @@ pub async fn create_rpc_connections(
                     let ipv4_url = format!("{}://127.0.0.1:{}", prefix, port);
                     let ipv6 = PeggyQueryClient::connect(ipv6_url.clone()).await;
                     let ipv4 = PeggyQueryClient::connect(ipv4_url.clone()).await;
+                    warn!("Trying fallback urls {} {}", ipv6_url, ipv4_url);
                     match (ipv4, ipv6) {
                         (Ok(v), Err(_)) => {
                             info!("Url fallback succeeded, your cosmos gRPC url {} has been corrected to {}", grpc_url, ipv4_url);
@@ -68,6 +78,10 @@ pub async fn create_rpc_connections(
                     let https_on_443_url = format!("https://{}:443", body);
                     let https_on_80 = PeggyQueryClient::connect(https_on_80_url.clone()).await;
                     let https_on_443 = PeggyQueryClient::connect(https_on_443_url.clone()).await;
+                    warn!(
+                        "Trying fallback urls {} {}",
+                        https_on_443_url, https_on_80_url
+                    );
                     match (https_on_80, https_on_443) {
                         (Ok(v), Err(_)) => {
                             info!("Https upgrade succeeded, your cosmos gRPC url {} has been corrected to {}", grpc_url, https_on_80_url);
@@ -98,6 +112,10 @@ pub async fn create_rpc_connections(
             Ok(_) => contact = Some(base_contact),
             // did not work, now we check if it's localhost
             Err(e) => {
+                warn!(
+                    "Failed to access Cosmos Leagcy RPC with {:?} trying fallback options",
+                    e
+                );
                 if legacy_rpc_url.to_lowercase().contains("localhost") {
                     let port = url.port().unwrap_or(80);
                     // this should be http or https
@@ -108,6 +126,7 @@ pub async fn create_rpc_connections(
                     let ipv4_contact = Contact::new(&ipv4_url, timeout);
                     let ipv6_test = ipv6_contact.get_latest_block().await;
                     let ipv4_test = ipv4_contact.get_latest_block().await;
+                    warn!("Trying fallback urls {} {}", ipv6_url, ipv4_url);
                     match (ipv4_test, ipv6_test) {
                         (Ok(_), Err(_)) => {
                             info!("Url fallback succeeded, your cosmos legacy rpc url {} has been corrected to {}", legacy_rpc_url, ipv4_url);
@@ -131,6 +150,10 @@ pub async fn create_rpc_connections(
                     let https_on_443_contact = Contact::new(&https_on_443_url, timeout);
                     let https_on_80_test = https_on_80_contact.get_latest_block().await;
                     let https_on_443_test = https_on_443_contact.get_latest_block().await;
+                    warn!(
+                        "Trying fallback urls {} {}",
+                        https_on_443_url, https_on_80_url
+                    );
                     match (https_on_80_test, https_on_443_test) {
                         (Ok(_), Err(_)) => {
                             info!("Https upgrade succeeded, your cosmos legacy rpc url {} has been corrected to {}", legacy_rpc_url, https_on_80_url);
@@ -154,13 +177,17 @@ pub async fn create_rpc_connections(
             .unwrap_or_else(|_| panic!("Invalid Ethereum RPC url {}", eth_rpc_url));
         check_scheme(&url, &eth_rpc_url);
         let eth_url = eth_rpc_url.trim_end_matches('/');
-        let base_web30 = Contact::new(&eth_rpc_url, timeout);
-        let try_base = base_web30.get_latest_block().await;
+        let base_web30 = Web3::new(&eth_url, timeout);
+        let try_base = base_web30.eth_block_number().await;
         match try_base {
             // it worked, lets go!
-            Ok(_) => contact = Some(base_web30),
+            Ok(_) => web3 = Some(base_web30),
             // did not work, now we check if it's localhost
             Err(e) => {
+                warn!(
+                    "Failed to access Ethereum RPC with {:?} trying fallback options",
+                    e
+                );
                 if eth_url.to_lowercase().contains("localhost") {
                     let port = url.port().unwrap_or(80);
                     // this should be http or https
@@ -171,6 +198,7 @@ pub async fn create_rpc_connections(
                     let ipv4_web3 = Web3::new(&ipv4_url, timeout);
                     let ipv6_test = ipv6_web3.eth_block_number().await;
                     let ipv4_test = ipv4_web3.eth_block_number().await;
+                    warn!("Trying fallback urls {} {}", ipv6_url, ipv4_url);
                     match (ipv4_test, ipv6_test) {
                         (Ok(_), Err(_)) => {
                             info!("Url fallback succeeded, your Ethereum rpc url {} has been corrected to {}", eth_rpc_url, ipv4_url);
@@ -194,6 +222,10 @@ pub async fn create_rpc_connections(
                     let https_on_443_web3 = Web3::new(&https_on_443_url, timeout);
                     let https_on_80_test = https_on_80_web3.eth_block_number().await;
                     let https_on_443_test = https_on_443_web3.eth_block_number().await;
+                    warn!(
+                        "Trying fallback urls {} {}",
+                        https_on_443_url, https_on_80_url
+                    );
                     match (https_on_80_test, https_on_443_test) {
                         (Ok(_), Err(_)) => {
                             info!("Https upgrade succeeded, your Ethereum rpc url {} has been corrected to {}", eth_rpc_url, https_on_80_url);
@@ -227,5 +259,79 @@ fn check_scheme(input: &Url, original_string: &str) {
             "Your url {} has an invalid scheme, please chose http or https",
             original_string
         )
+    }
+}
+
+/// This function checks the orchestrator delegate addresses
+/// for consistency what this means is that it takes the Ethereum
+/// address and Orchestrator address from the Orchestrator and checks
+/// that both are registered and internally consistent.
+pub async fn check_delegate_addresses(
+    client: &mut PeggyQueryClient<Channel>,
+    delegate_eth_address: EthAddress,
+    delegate_orchestrator_address: CosmosAddress,
+) {
+    let eth_response = client
+        .get_delegate_key_by_eth(QueryDelegateKeysByEthAddress {
+            eth_address: delegate_eth_address.to_string(),
+        })
+        .await;
+    let orchestrator_response = client
+        .get_delegate_key_by_orchestrator(QueryDelegateKeysByOrchestratorAddress {
+            orchestrator_address: delegate_orchestrator_address.to_string(),
+        })
+        .await;
+    info!("{:?} {:?}", eth_response, orchestrator_response);
+    match (eth_response, orchestrator_response) {
+        (Ok(e), Ok(o)) => {
+            let e = e.into_inner();
+            let o = o.into_inner();
+            let req_delegate_orchestrator_address: CosmosAddress =
+                e.orchestrator_address.parse().unwrap();
+            let req_delegate_eth_address: EthAddress = o.eth_address.parse().unwrap();
+            if req_delegate_eth_address != delegate_eth_address
+                && req_delegate_orchestrator_address != delegate_orchestrator_address
+            {
+                error!("Your Delegate Ethereum and Orchestrator addresses are both incorrect!");
+                error!(
+                    "You provided {}  Correct Value {}",
+                    delegate_eth_address, req_delegate_eth_address
+                );
+                error!(
+                    "You provided {}  Correct Value {}",
+                    delegate_orchestrator_address, req_delegate_orchestrator_address
+                );
+                error!("In order to resolve this issue you should double check your input value or re-register your delegate keys");
+                exit(1);
+            } else if req_delegate_eth_address != delegate_eth_address {
+                error!("Your Delegate Ethereum address is incorrect!");
+                error!(
+                    "You provided {}  Correct Value {}",
+                    delegate_eth_address, req_delegate_eth_address
+                );
+                error!("In order to resolve this issue you should double check how you input your eth private key");
+                exit(1);
+            } else if req_delegate_orchestrator_address != delegate_orchestrator_address {
+                error!("Your Delegate Orchestrator address is incorrect!");
+                error!(
+                    "You provided {}  Correct Value {}",
+                    delegate_eth_address, req_delegate_eth_address
+                );
+                error!("In order to resolve this issue you should double check how you input your Orchestrator address phrase, make sure you didn't use your Validator phrase!");
+                exit(1);
+            }
+
+            if e.validator_address != o.validator_address {
+                error!("You are using delegate keys from two different validator addresses!");
+                error!("If you get this error message I would just blow everything away and start again");
+                exit(1);
+            }
+        }
+        (Err(_), Ok(_)) | (Ok(_), Err(_)) => {
+            panic!("Failed to check delegate Eth address. Maybe try running the program again? If that doesn't work try registering delegate keys again")
+        }
+        (Err(_), Err(_)) => {
+            panic!("Delegate addresses are not set! Please Register your delegate keys and make sure your Althea binary is updated!")
+        }
     }
 }
